# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/services/maturity-index:
    get:
      operationId: ServicesController_maturityIndex
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaturityIndexResponse'
      security:
      - bearer: []
      summary: Service maturity index with scoring and domain breakdown
      tags:
      - Services
  /api/v1/services/maturity-index/mermaid:
    get:
      operationId: ServicesController_maturityIndexMermaid
      parameters: []
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: Maturity index as a Mermaid flowchart diagram
      tags:
      - Services
  /api/v1/services:
    get:
      operationId: ServicesController_list
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
      security:
      - bearer: []
      summary: List all services
      tags:
      - Services
  /api/v1/services/{id}:
    get:
      operationId: ServicesController_getById
      parameters:
      - name: id
        required: true
        in: path
        description: 'Service ID (with or without svc: prefix)'
        schema:
          example: svc:wafed
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailResponse'
        '404':
          description: Service not found
      security:
      - bearer: []
      summary: Get service detail with screens, roles, integrations, dependencies,
        documents
      tags:
      - Services
  /api/v1/services/{id}/dependencies:
    get:
      operationId: ServicesController_getDependencies
      parameters:
      - name: id
        required: true
        in: path
        schema:
          example: svc:wafed
          type: string
      responses:
        '200':
          description: Dependencies returned
        '404':
          description: Service not found
      security:
      - bearer: []
      summary: Get service dependencies (dependsOn + dependedBy)
      tags:
      - Services
  /api/v1/services/{id}/impact-path:
    get:
      operationId: ServicesController_getImpactPath
      parameters:
      - name: id
        required: true
        in: path
        schema:
          example: svc:wafed
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImpactPathResponse'
        '404':
          description: Service not found
      security:
      - bearer: []
      summary: Get impact path — screens, fields, actions, and integrations for a
        service
      tags:
      - Services
  /api/v1/mermaid/service/{id}:
    get:
      operationId: MermaidController_serviceArchitecture
      parameters:
      - name: id
        required: true
        in: path
        schema:
          example: svc:wafed
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: Generate Mermaid architecture diagram for a service
      tags:
      - Mermaid
  /api/v1/mermaid/dependencies/{id}:
    get:
      operationId: MermaidController_dependencies
      parameters:
      - name: id
        required: true
        in: path
        schema:
          example: svc:wafed
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: Generate Mermaid dependency diagram for a service
      tags:
      - Mermaid
  /api/v1/mermaid/search:
    get:
      operationId: MermaidController_search
      parameters:
      - name: q
        required: true
        in: query
        schema:
          example: ترقيات
          type: string
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: Search and visualize matching nodes as Mermaid diagram
      tags:
      - Mermaid
  /api/v1/mermaid/analyze:
    post:
      operationId: MermaidController_analyze
      parameters: []
      responses:
        '201':
          description: ''
      security:
      - bearer: []
      summary: Analyze change impact and return Mermaid diagram
      tags:
      - Mermaid
  /api/v1/mermaid/explore:
    get:
      operationId: MermaidController_explore
      parameters: []
      responses:
        '200':
          description: ''
      security:
      - bearer: []
      summary: Full MASAR service graph as Mermaid diagram
      tags:
      - Mermaid
...
info:
  title: MASAR Impact Analysis API
  description: API for impact analysis of MASAR government services. Includes service
    graph traversal, async job processing, and Mermaid diagram generation.
  version: 2.0.0
  contact: {}
...
servers:
- url: https://myapi.businessanalystcrew.org
  description: Production
- url: http://localhost:3000
  description: Local
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: API token from system administrator
  schemas:
...
    MaturityEntry:
      type: object
      properties:
        id:
          type: string
        name_ar:
          type: string
        name_en:
          type: string
        domain:
          type: string
        maturity:
          type: string
          example: L4
        impact_readiness:
          type: string
          example: IR-4
        maturity_score:
          type: number
          example: 4
        readiness_score:
          type: number
          example: 4
        maturity_index:
          type: number
          example: 4
        tier:
          type: string
          example: متقدم
      required:
      - id
      - name_ar
      - name_en
      - domain
      - maturity
      - impact_readiness
      - maturity_score
      - readiness_score
      - maturity_index
      - tier
    MaturityIndexResponse:
      type: object
      properties:
        overall_index:
          type: number
          example: 3.04
        tier:
          type: string
          example: متوسط
        total_services:
          type: number
          example: 25
        distribution:
          type: object
          example:
            L4: 6
            L3: 14
            L2: 5
        readiness:
          type: object
          example:
            IR-4: 2
            IR-3: 22
            IR-2: 1
        by_domain:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/MaturityByDomain'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/MaturityEntry'
      required:
      - overall_index
      - tier
      - total_services
      - distribution
      - readiness
      - by_domain
      - entries
    ServiceListItem:
      type: object
      properties:
        id:
          type: string
          example: svc:wafed
        name_ar:
          type: string
          example: تأييد التعاقد لغير السعوديين - وافد
        name_en:
          type: string
          example: Contract Endorsement for Non-Saudis - WAFED
        description:
          type: object
        domain:
          type: string
          example: Employment
        maturity:
          type: string
          example: L4
        impact_readiness:
          type: string
          example: IR-4
      required:
      - id
      - name_ar
      - name_en
      - domain
      - maturity
      - impact_readiness
    ServiceListResponse:
      type: object
      properties:
        count:
          type: number
          example: 25
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceListItem'
      required:
      - count
      - services
    ScreenDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: تقديم طلب تعاقد
        type:
          type: string
          example: FORM
      required:
      - id
      - name
      - type
    RoleDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: مختص جهة حكومية
        scope:
          type: string
          example: Government
      required:
      - id
      - name
      - scope
    IntegrationDto:
      type: object
      properties:
        name:
          type: string
          example: JICS Integration - WAFED
        direction:
          type: string
          example: BIDIRECTIONAL
        systems:
          example:
          - النظام المركزي
          type: array
          items:
            type: string
      required:
      - name
      - direction
      - systems
    DependencyRef:
      type: object
      properties:
        id:
          type: string
          example: svc:jics
        name:
          type: string
          example: النظام المركزي
      required:
      - id
      - name
    DependenciesDto:
      type: object
      properties:
        dependsOn:
          type: array
          items:
            $ref: '#/components/schemas/DependencyRef'
        dependedBy:
          type: array
          items:
            $ref: '#/components/schemas/DependencyRef'
      required:
      - dependsOn
      - dependedBy
    DocumentRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: وثيقة معمارية الحل - وافد
        version:
          type: string
          example: '1.0'
      required:
      - id
      - name
      - version
    ServiceDetailResponse:
      type: object
      properties:
        service:
          type: object
        screens:
          type: array
          items:
            $ref: '#/components/schemas/ScreenDto'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleDto'
        integrations:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationDto'
        dependencies:
          $ref: '#/components/schemas/DependenciesDto'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/DocumentRef'
      required:
      - service
      - screens
      - roles
      - integrations
      - dependencies
      - documents
    ImpactPathResponse:
      type: object
      properties:
        service:
          type: object
        path:
          type: object
      required:
      - service
      - path
...